requirements.txt
psychopy
screeninfo
pyserial
psychtoolbox
pyyaml
numpy